Replace ac_int for standard C++ types.
git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@3510 86035c5f-d46c-4136-a2d5-fe63f1d4a782

+++ branches/ud/hw/framework/catapult.h
@@ -22,7 +22,7 @@ private:
 
 public:
     // TODO: Payload size should be Traits<System>::pkt_size
-    unsigned int payload;
+    unsigned long payload;
     Header header;
     // Source address for received messages, destination address for transmitted
     // messages
